Fix #420 - grass on yavin had corrupted colours.
This commit fixes a problem with corrupted colours on quick sprites,
when running as a 64-bit binary. unsigned longs were used for 4-byte
colours. While this would work on 32-bit systems, an unsigned long has 8
bytes on 64-bit systems. Colour types have been changed to uint32_t to
ensure the correct size on all systems.

The same has also been done for the GL_State function, which previously
accepted a bitfield of states as an unsigned long. Now it requires state
to be a uint32_t.

+++ code/rd-vanilla/tr_quicksprite.cpp
@@ -151,7 +151,7 @@ void CQuickSpriteSystem::Flush(void)
 }
 
 
-void CQuickSpriteSystem::StartGroup(textureBundle_t *bundle, unsigned long glbits, int fogIndex )
+void CQuickSpriteSystem::StartGroup(textureBundle_t *bundle, uint32_t glbits, int fogIndex )
 {
 	mNextVert = 0;
 
@@ -200,7 +200,7 @@ void CQuickSpriteSystem::Add(float *pointdata, color4ub_t color, vec2_t fog)
 {
 	float *curcoord;
 	float *curfogtexcoord;
-	unsigned long *curcolor;
+	uint32_t *curcolor;
 
 	if (mNextVert>SHADER_MAX_VERTEXES-4)
 	{
@@ -212,10 +212,10 @@ void CQuickSpriteSystem::Add(float *pointdata, color4ub_t color, vec2_t fog)
 
 	// Set up color
 	curcolor = &mColors[mNextVert];
-	*curcolor++ = *(unsigned long *)color;
-	*curcolor++ = *(unsigned long *)color;
-	*curcolor++ = *(unsigned long *)color;
-	*curcolor++ = *(unsigned long *)color;
+	*curcolor++ = *(uint32_t *)color;
+	*curcolor++ = *(uint32_t *)color;
+	*curcolor++ = *(uint32_t *)color;
+	*curcolor++ = *(uint32_t *)color;
 
 	if (fog)
 	{
